Rename and move PipelineJobExecutor
